Vcp 6 - backend

+++ clinic/src/api/HealthPackageAPI.js
@@ -1,3 +1,4 @@
+
 import HealthPackageService from '../service/health-package-service.js';
 import {
 	EMPTY_SIZE,
@@ -5,6 +6,7 @@ import {
 	NOT_FOUND_STATUS_CODE,
 	OK_STATUS_CODE,
 } from '../utils/Constants.js';
+import { isValidMongoId } from '../utils/Validation.js';
 
 export const healthPackage = (app) => {
 	const service = new HealthPackageService();
@@ -42,10 +44,25 @@ export const healthPackage = (app) => {
 		}
 	});
 
+	app.patch('/package/:id', async (req,res)=>{
+        const updateData = req.body;
+        const id = req.params.id;
+		if (!isValidMongoId(id))
+			return res.status(ERROR_STATUS_CODE).json({ message: 'Invalid ID' });
+        try{
+            const updatedPackage = await service.updatePackage(id, updateData);
+            res.status(OK_STATUS_CODE).json({updatedPackage});
+        }catch(err){
+            res.status(ERROR_STATUS_CODE).json({err : err.message});
+        }
+    });
+
 	app.delete('/packages/:id', async (req, res) => {
 		const id = req.params.id;
+		if (!isValidMongoId(id))
+			return res.status(ERROR_STATUS_CODE).json({ message: 'Invalid ID' });
+
 		try {
-			console.log(id);
 			const deletedPackage = await service.deletePackage(id);
 			res.status(OK_STATUS_CODE).json({ deletedPackage });
 		} catch (err) {
